+++ ios/Classes/WallcraftManagerPlugin.h
@@ -1,0 +1,5 @@
+# ios/Classes/WallcraftManagerPlugin.h
+#import <Flutter/Flutter.h>
+
+@interface WallcraftManagerPlugin : NSObject<FlutterPlugin>
+@end